move the test for trasform function into its own file.
