+++ tests/unit/transformer/test_node_transformer.py
@@ -7,11 +7,10 @@ import sys
 
 import asttokens
 import pytest
-from mock import MagicMock, patch
+from mock import MagicMock
 
 from lineapy.transformer.node_transformer import NodeTransformer
 from lineapy.transformer.source_giver import SourceGiver
-from lineapy.transformer.transform_code import transform
 
 
 def _get_ast_node(code):
@@ -21,21 +20,6 @@ def _get_ast_node(code):
         SourceGiver().transform(node)
 
     return node
-
-
-@patch(
-    "lineapy.transformer.transform_code.NodeTransformer",
-)
-def test_transform_fn(nt_mock: MagicMock):
-    """
-    Test that the transform function calls the NodeTransformer
-    """
-    mocked_tracer = MagicMock()
-    source_location = MagicMock()
-    transform("x = 1", source_location, mocked_tracer)
-    nt_mock.assert_called_once_with("x = 1", source_location, mocked_tracer)
-    mocked_tracer.db.commit.assert_called_once()
-    # TODO - test that source giver is called only for 3.7 and below
 
 
 class TestNodeTransformer:
